implemented worker call

+++ app/src/main/java/il/co/rootsapp/RootItem.kt
@@ -4,10 +4,10 @@ import java.io.Serializable
 import java.time.LocalDateTime
 
 data class RootItem(
-    val num: Int,
-    val lowerBound: Int = 0,
-    val root1: Int? = null,
-    val root2: Int? = null,
+    val num: Long,
+    val lowerBound: Long = 0,
+    val root1: Long? = null,
+    val root2: Long? = null,
     var isDone: Boolean = false,
     private val creationDT: LocalDateTime = LocalDateTime.now()
 ) : Serializable, Comparable<RootItem> {
